Create AdminUserModal that handles both creating and editing users
